backend/requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.12

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0

# Redis
redis>=5.0.0

# Task Queue
celery[redis]>=5.4.0

# Browser Automation
playwright>=1.49.0

# HTTP Client
httpx[socks]>=0.28.0
h2>=4.0.0
curl_cffi>=0.7.0

# Content Extraction
trafilatura>=2.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
markdownify>=0.14.0

# LLM
litellm>=1.55.0

# Auth & Security
pyjwt>=2.9.0
bcrypt>=4.0.0
cryptography>=44.0.0

# Validation
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Robots.txt
robotexclusionrulesparser>=1.7.1

# Search
ddgs>=9.0.0

# Metrics
prometheus-client>=0.21.0

# Utilities
python-dotenv>=1.0.0
uuid6>=2024.7.10

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
